1. addressing PR comments, 2. rename files to remove - (hyphen)
