+++ blocks/page_list/templates/itinerary/ItineraryStaticData.js
