+++ blocks/page_list/templates/itinerary/ItineraryStore.js
@@ -1,37 +1,37 @@
-import { dispatch, register } from './Itinerary-Dispatcher';
+import { dispatch, register } from './ItineraryDispatcher';
 import { EventEmitter } from 'events';
-import ItineraryConstants from './Itinerary-Constants';
-import walks from './Itinerary-Static-Data';
+import ItineraryConstants from './ItineraryConstants';
+import walks from './ItineraryStaticData';
 
 const CHANGE_EVENT = 'change';
 
-let _walks = {walks:walks.slice(0,-1)}; //grabbing static walks for now
+let _walksList = {walks: walks.slice(0, -1)}; //grabbing static walks for now
 
 const _removeWalk = ( id ) => {
-  _walks.splice( _walks.findIndex(walks => walks.id === id),1);
+  _walksList.splice( _walksList.findIndex(walks => walks.id === id),1);
 };
 
 
 const _addWalk = ( id ) => {
-  let walkExists = _walks.walks.findIndex(walk => walk.id === id);
+  let walkExists = _walksList.walks.findIndex(walk => walk.id === id);
 
   if(walkExists === -1) {
-    let walkToAdd = walks.filter(walk => walk.id === id)[0];
-    _walks.unshift(walkToAdd);
+    let walkToAdd = walks.find(walk => walk.id === id)[0];
+    _walksList.unshift(walkToAdd);
   }else{
     console.log('Walk already exists, to notify the user');
   }
 };
 
-//walks received from API used to update _walks
+//walks received from API used to update _walksList
 const _updateWalks = ( walks ) => {
-  _walks = walks.slice();
+  _walksList = walks.slice();
 };
 
 const ItineraryStore = Object.assign(EventEmitter.prototype, {
   emitChange() {
     this.emit( CHANGE_EVENT );
-		},
+  },
 
   addChangeListener( callback ){
     this.on( CHANGE_EVENT, callback);
@@ -42,7 +42,7 @@ const ItineraryStore = Object.assign(EventEmitter.prototype, {
   },
 
   getItinerary(){
-    return _walks;
+    return _walksList;
   },
 
   //TODO: use _updateWalks to receive walks from server via API call
