+++ blocks/page_list/templates/itinerary/view.less
@@ -1,24 +1,30 @@
-.itinerary-walk {
-  max-width: 70%;
-  float: left;
-}
+.itinerary {
 
-.itinerary-remove-walk {
-  float: right;
-  max-width: 30%;
-  position: relative;
-  transform: translateY(100%);
-}
+  .walk {
+    .clearfix();
+    max-width: 70%;
+    float: left;
+  }
 
-.itinerary-remove-button {
-  border: none;
-  background-color: transparent;
-  color: #959595;
-  min-width: 100px;
-  outline: none;
-  text-align: right;
+  .remove-walk {
+    float: right;
+    max-width: 30%;
+    position: relative;
+    transform: translateY(100%);
+  }
 
-  &:hover {
-    color: black;
+  .remove-button {
+    border: none;
+    background-color: transparent;
+    color: #959595;
+    min-width: 100px;
+    outline: none;
+    text-align: right;
+
+    &:hover {
+      color: black;
+    }
   }
 }
+
+
